Add files via upload
BlockandGridBuy.cs is a script to be used with Crunch Econ V3 for The Habitat. It allows the station to use custom data of button blocks and allows customization to be able to purchase blocks on a grid or purchase an entire pre-made grid.

Grids - heavy miner transport.sbc - Little Runner - Respawn Shuttle - Are designed to be used with the station unless you decide to use your own.